Refactor filter testcases to sepertae class
